KAFKA-14239: Merge StateRestorationIntegrationTest into RestoreIntegrationTest (#12670)
This PR makes the following changes:

* Moves the only test in StateRestorationIntegrationTest into RestoreIntegrationTest
* Deletes StateRestorationIntegrationTest

Reviewers: Guozhang Wang <wangguoz@gmail.com>
